Add Situation value object
There is repeated logic throughout the services which verifies
the validity of the situation string for `ActionFeedback`
(limited to 100 chars). This class moves that validation
to a single place but `ActionFeedback` still expects a string.
For now the `toString()` method can be used but it would be
better if `ActionFeedback` accepted a proper `Situation` type
in the future (not sure if this is possible with the jaxb
generated class).